requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses==0.6
Django==5.0.1
django-extensions==3.2.3
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pillow==10.2.0
psycopg2-binary==2.9.9
pyderive3==0.0.6
PyJWT==2.8.0
pypool3==0.0.2
pyserve3==0.0.7
PySocks==1.7.1
pystructs3==0.0.7
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
openai==v0.27.0
tqdm==4.62.3
python-dotenv==0.19.0
tiktoken==0.6.0
urllib3==2.2.0
wsproto==1.2.0

